improve CCM/CSI migration logic
