+++ pkg/resources/cloudcontroller/anexia.go
@@ -21,12 +21,12 @@ import (
 
 	"k8c.io/kubermatic/v2/pkg/resources"
 	"k8c.io/kubermatic/v2/pkg/resources/reconciling"
-	"k8c.io/kubermatic/v2/pkg/resources/vpnsidecar"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
+	"k8s.io/utils/pointer"
 )
 
 const AnexiaCCMDeploymentName = "anx-cloud-controller-manager"
@@ -50,19 +50,8 @@ func anexiaDeploymentCreator(data *resources.TemplateData) reconciling.NamedDepl
 				Labels: podLabels,
 			}
 
-			f := false
-			deployment.Spec.Template.Spec.AutomountServiceAccountToken = &f
-
-			deployment.Spec.Template.Spec.Volumes = append(getVolumes(data.IsKonnectivityEnabled()),
-				corev1.Volume{
-					Name: resources.CloudConfigSeedSecretName,
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: resources.CloudConfigSeedSecretName,
-						},
-					},
-				})
-
+			deployment.Spec.Template.Spec.AutomountServiceAccountToken = pointer.Bool(false)
+			deployment.Spec.Template.Spec.Volumes = getVolumes(data.IsKonnectivityEnabled(), true)
 			deployment.Spec.Template.Spec.Containers = []corev1.Container{
 				{
 					Name:  ccmContainerName,
@@ -74,8 +63,9 @@ func anexiaDeploymentCreator(data *resources.TemplateData) reconciling.NamedDepl
 						fmt.Sprintf("--cluster-name=%s", data.Cluster().Spec.HumanReadableName),
 						"--kubeconfig=/etc/kubernetes/kubeconfig/kubeconfig",
 					},
-					Env: []corev1.EnvVar{
-						{
+					Env: append(
+						getEnvVars(),
+						corev1.EnvVar{
 							Name: "ANEXIA_TOKEN",
 							ValueFrom: &corev1.EnvVarSource{
 								SecretKeyRef: &corev1.SecretKeySelector{
@@ -86,11 +76,12 @@ func anexiaDeploymentCreator(data *resources.TemplateData) reconciling.NamedDepl
 								},
 							},
 						},
-						{
+						corev1.EnvVar{
 							Name:  "ANEXIA_AUTO_DISCOVER_LOAD_BALANCER",
 							Value: "true",
 						},
-					},
+					),
+					VolumeMounts: getVolumeMounts(true),
 					Ports: []corev1.ContainerPort{
 						{
 							Name:          "http",
@@ -126,25 +117,9 @@ func anexiaDeploymentCreator(data *resources.TemplateData) reconciling.NamedDepl
 						SuccessThreshold:    1,
 						FailureThreshold:    3,
 					},
-					VolumeMounts: append(getVolumeMounts(), corev1.VolumeMount{
-						Name:      resources.CloudConfigSeedSecretName,
-						MountPath: "/etc/kubernetes/cloud",
-						ReadOnly:  true,
-					}),
 				},
 			}
 
-			if !data.IsKonnectivityEnabled() {
-				openvpnSidecar, err := vpnsidecar.OpenVPNSidecarContainer(data, openvpnClientContainerName)
-				if err != nil {
-					return nil, fmt.Errorf("failed to get openvpn sidecar: %w", err)
-				}
-				deployment.Spec.Template.Spec.Containers = append(deployment.Spec.Template.Spec.Containers, *openvpnSidecar)
-			}
-
-			if err != nil {
-				return nil, err
-			}
 			return deployment, nil
 		}
 	}
