+++ pkg/resources/cloudcontroller/azure.go
@@ -22,7 +22,6 @@ import (
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
 	"k8c.io/kubermatic/v2/pkg/resources"
 	"k8c.io/kubermatic/v2/pkg/resources/reconciling"
-	"k8c.io/kubermatic/v2/pkg/resources/vpnsidecar"
 	"k8c.io/kubermatic/v2/pkg/semver"
 
 	appsv1 "k8s.io/api/apps/v1"
@@ -76,36 +75,21 @@ func azureDeploymentCreator(data *resources.TemplateData) reconciling.NamedDeplo
 				return nil, err
 			}
 
-			dep.Spec.Template.Spec.AutomountServiceAccountToken = pointer.BoolPtr(false)
-
 			version, err := getAzureVersion(data.Cluster().Status.Versions.ControlPlane)
 			if err != nil {
 				return nil, err
 			}
 
-			dep.Spec.Template.Spec.Volumes = append(getVolumes(data.IsKonnectivityEnabled()),
-				corev1.Volume{
-					Name: resources.CloudConfigSeedSecretName,
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: resources.CloudConfigSeedSecretName,
-						},
-					},
-				})
-
+			dep.Spec.Template.Spec.AutomountServiceAccountToken = pointer.BoolPtr(false)
+			dep.Spec.Template.Spec.Volumes = getVolumes(data.IsKonnectivityEnabled(), true)
 			dep.Spec.Template.Spec.Containers = []corev1.Container{
 				{
-					Name:    ccmContainerName,
-					Image:   data.ImageRegistry(resources.RegistryMCR) + "/oss/kubernetes/azure-cloud-controller-manager:v" + version,
-					Command: []string{"cloud-controller-manager"},
-					Args:    getAzureFlags(data),
-					VolumeMounts: append(getVolumeMounts(),
-						corev1.VolumeMount{
-							Name:      resources.CloudConfigSeedSecretName,
-							MountPath: "/etc/kubernetes/cloud",
-							ReadOnly:  true,
-						},
-					),
+					Name:         ccmContainerName,
+					Image:        data.ImageRegistry(resources.RegistryMCR) + "/oss/kubernetes/azure-cloud-controller-manager:v" + version,
+					Command:      []string{"cloud-controller-manager"},
+					Args:         getAzureFlags(data),
+					Env:          getEnvVars(),
+					VolumeMounts: getVolumeMounts(true),
 					LivenessProbe: &corev1.Probe{
 						ProbeHandler: corev1.ProbeHandler{
 							HTTPGet: &corev1.HTTPGetAction{
@@ -127,19 +111,11 @@ func azureDeploymentCreator(data *resources.TemplateData) reconciling.NamedDeplo
 				ccmContainerName: azureResourceRequirements.DeepCopy(),
 			}
 
-			if !data.IsKonnectivityEnabled() {
-				openvpnSidecar, err := vpnsidecar.OpenVPNSidecarContainer(data, openvpnClientContainerName)
-				if err != nil {
-					return nil, fmt.Errorf("failed to get openvpn sidecar: %w", err)
-				}
-				dep.Spec.Template.Spec.Containers = append(dep.Spec.Template.Spec.Containers, *openvpnSidecar)
-				defResourceRequirements[openvpnSidecar.Name] = openvpnSidecar.Resources.DeepCopy()
-			}
-
 			err = resources.SetResourceRequirements(dep.Spec.Template.Spec.Containers, defResourceRequirements, nil, dep.Annotations)
 			if err != nil {
 				return nil, fmt.Errorf("failed to set resource requirements: %w", err)
 			}
+
 			return dep, nil
 		}
 	}
@@ -180,13 +156,4 @@ func getAzureFlags(data *resources.TemplateData) []string {
 		flags = append(flags, "--cluster-name", data.Cluster().Name)
 	}
 	return flags
-}
-
-// AzureCloudControllerSupported checks if this version of Kubernetes is supported
-// by our implementation of the external cloud controller.
-func AzureCloudControllerSupported(version semver.Semver) bool {
-	if _, err := getAzureVersion(version); err != nil {
-		return false
-	}
-	return true
 }
