+++ pkg/resources/cloudcontroller/util.go
@@ -17,11 +17,7 @@ limitations under the License.
 package cloudcontroller
 
 import (
-	"net/url"
-
-	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
 	"k8c.io/kubermatic/v2/pkg/resources"
-	"k8c.io/kubermatic/v2/pkg/version"
 
 	corev1 "k8s.io/api/core/v1"
 )
@@ -35,122 +31,7 @@ const (
 	v1240 = "1.24.0"
 )
 
-// ExternalCloudControllerFeatureSupported checks if the cloud provider supports
-// external CCM.
-func ExternalCloudControllerFeatureSupported(dc *kubermaticv1.Datacenter, cluster *kubermaticv1.Cluster, incompatibilities ...*version.ProviderIncompatibility) bool {
-	// This function is called during cluster creation and at that time, the
-	// cluster status might not have been initially set yet, so we must ensure
-	// to fallback to the spec version.
-	v := cluster.Status.Versions.ControlPlane
-	if v == "" {
-		v = cluster.Spec.Version
-	}
-
-	switch {
-	case cluster.Spec.Cloud.Openstack != nil:
-		// When using OpenStack external CCM with Open Telekom Cloud the creation
-		// of LBs fail as documented in the issue below:
-		// https://github.com/kubernetes/cloud-provider-openstack/issues/960
-		// Falling back to the in-tree CloudProvider mitigates the problem, even if
-		// not all features are expected to work properly (e.g.
-		// `manage-security-groups` should be set to false in cloud config).
-		//
-		// TODO This is a dirty hack to temporarily support OTC using
-		// Openstack provider, remove this when dedicated OTC support is
-		// introduced in Kubermatic.
-		return !isOTC(dc.Spec.Openstack) && OpenStackCloudControllerSupported(v)
-
-	case cluster.Spec.Cloud.Hetzner != nil:
-		return cluster.Spec.Cloud.Hetzner.Network != "" || dc.Spec.Hetzner.Network != ""
-
-	case cluster.Spec.Cloud.VSphere != nil:
-		supported, err := version.IsSupported(v.Semver(), kubermaticv1.VSphereCloudProvider, incompatibilities, kubermaticv1.ExternalCloudProviderCondition)
-		if err != nil {
-			return false
-		}
-		return supported
-
-	case cluster.Spec.Cloud.Anexia != nil:
-		return true
-
-	case cluster.Spec.Cloud.Kubevirt != nil:
-		return true
-
-	case cluster.Spec.Cloud.Azure != nil:
-		return AzureCloudControllerSupported(v)
-
-	default:
-		return false
-	}
-}
-
-// MigrationToExternalCloudControllerSupported checks if the cloud provider supports the migration to the
-// external CCM.
-func MigrationToExternalCloudControllerSupported(dc *kubermaticv1.Datacenter, cluster *kubermaticv1.Cluster, incompatibilities ...*version.ProviderIncompatibility) bool {
-	// This function is called during cluster creation and at that time, the
-	// cluster status might not have been initially set yet, so we must ensure
-	// to fallback to the spec version.
-	v := cluster.Status.Versions.ControlPlane
-	if v == "" {
-		v = cluster.Spec.Version
-	}
-	switch {
-	case cluster.Spec.Cloud.Openstack != nil:
-		// When using OpenStack external CCM with Open Telekom Cloud the creation
-		// of LBs fail as documented in the issue below:
-		// https://github.com/kubernetes/cloud-provider-openstack/issues/960
-		// Falling back to the in-tree CloudProvider mitigates the problem, even if
-		// not all features are expected to work properly (e.g.
-		// `manage-security-groups` should be set to false in cloud config).
-		//
-		// TODO This is a dirty hack to temporarily support OTC using
-		// Openstack provider, remove this when dedicated OTC support is
-		// introduced in Kubermatic.
-		return !isOTC(dc.Spec.Openstack) && OpenStackCloudControllerSupported(v)
-
-	case cluster.Spec.Cloud.VSphere != nil:
-		supported, err := version.IsSupported(v.Semver(), kubermaticv1.VSphereCloudProvider, incompatibilities, kubermaticv1.ExternalCloudProviderCondition)
-		if err != nil {
-			return false
-		}
-		return supported
-
-	case cluster.Spec.Cloud.Azure != nil:
-		versionSupported, err := version.IsSupported(v.Semver(), kubermaticv1.AzureCloudProvider, incompatibilities, kubermaticv1.ExternalCloudProviderCondition)
-		if err != nil {
-			return false
-		}
-		return versionSupported && AzureCloudControllerSupported(v)
-
-	default:
-		return false
-	}
-}
-
-// ExternalCloudControllerClusterName checks if the ClusterFeatureCCMClusterName is supported
-// for the cloud provider.
-func ExternalCloudControllerClusterName(cluster *kubermaticv1.Cluster) bool {
-	switch {
-	case cluster.Spec.Cloud.Openstack != nil:
-		return true
-	case cluster.Spec.Cloud.Azure != nil:
-		return true
-	default:
-		return false
-	}
-}
-
-// isOTC returns `true` if the OpenStack Datacenter uses OTC (i.e.
-// Open Telekom Cloud), `false` otherwise.
-func isOTC(dc *kubermaticv1.DatacenterSpecOpenstack) bool {
-	u, err := url.Parse(dc.AuthURL)
-	if err != nil {
-		return false
-	}
-	return u.Host == "iam.eu-de.otc.t-systems.com"
-}
-
-func getVolumes(isKonnectivityEnabled bool) []corev1.Volume {
+func getVolumes(isKonnectivityEnabled bool, mountCloudConfig bool) []corev1.Volume {
 	vs := []corev1.Volume{
 		{
 			Name: resources.CloudControllerManagerKubeconfigSecretName,
@@ -160,6 +41,26 @@ func getVolumes(isKonnectivityEnabled bool) []corev1.Volume {
 				},
 			},
 		},
+		{
+			Name: resources.CABundleConfigMapName,
+			VolumeSource: corev1.VolumeSource{
+				ConfigMap: &corev1.ConfigMapVolumeSource{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: resources.CABundleConfigMapName,
+					},
+				},
+			},
+		},
+	}
+	if mountCloudConfig {
+		vs = append(vs, corev1.Volume{
+			Name: resources.CloudConfigSeedSecretName,
+			VolumeSource: corev1.VolumeSource{
+				Secret: &corev1.SecretVolumeSource{
+					SecretName: resources.CloudConfigSeedSecretName,
+				},
+			},
+		})
 	}
 	if !isKonnectivityEnabled {
 		vs = append(vs, corev1.Volume{
@@ -174,12 +75,36 @@ func getVolumes(isKonnectivityEnabled bool) []corev1.Volume {
 	return vs
 }
 
-func getVolumeMounts() []corev1.VolumeMount {
-	return []corev1.VolumeMount{
+func getVolumeMounts(mountCloudConfig bool) []corev1.VolumeMount {
+	mounts := []corev1.VolumeMount{
 		{
 			Name:      resources.CloudControllerManagerKubeconfigSecretName,
 			MountPath: "/etc/kubernetes/kubeconfig",
 			ReadOnly:  true,
+		},
+		{
+			Name:      resources.CABundleConfigMapName,
+			MountPath: "/etc/kubermatic/certs",
+			ReadOnly:  true,
+		},
+	}
+
+	if mountCloudConfig {
+		mounts = append(mounts, corev1.VolumeMount{
+			Name:      resources.CloudConfigSeedSecretName,
+			MountPath: "/etc/kubernetes/cloud",
+			ReadOnly:  true,
+		})
+	}
+
+	return mounts
+}
+
+func getEnvVars() []corev1.EnvVar {
+	return []corev1.EnvVar{
+		{
+			Name:  "SSL_CERT_FILE",
+			Value: "/etc/kubermatic/certs/ca-bundle.pem",
 		},
 	}
 }
